DATACMNS-1011 - Improved documentation of projections
Added an example using nested projections.

Clarified the criteria that make a projection a closed projection.

Removed any mentioning the option of changing what entities have a repository, because that decision should only be driven by fundamental design considerations.

Added a paragraph about when projections might be the wrong tool for the job and what alternatives might be better suited.